# Use a bufio.Reader for input.